# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "하드웨어 기반 라이선스 관리 서버",
        "title": "Studio License Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/licenses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "라이선스 목록을 페이징하여 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "관리자 - 라이선스"
                ],
                "summary": "라이선스 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "페이지 크기",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "상태 필터 (active, expired, revoked)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "검색어 (라이선스 키, 고객명, 이메일)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "조회 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.License"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "인증 필요",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "새로운 라이선스를 생성합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "관리자 - 라이선스"
                ],
                "summary": "라이선스 생성",
                "parameters": [
                    {
                        "description": "라이선스 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateLicenseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "생성 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.License"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "인증 필요",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/licenses/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 라이선스의 상세 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "관리자 - 라이선스"
                ],
                "summary": "라이선스 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "라이선스 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "조회 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.License"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "인증 필요",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "라이선스 없음",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "success, error",
                    "type": "string"
                }
            }
        },
...
        "models.CreateLicenseRequest": {
            "type": "object",
            "required": [
                "customer_email",
                "customer_name",
                "expires_at",
                "max_devices",
                "product_name"
            ],
            "properties": {
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "max_devices": {
                    "type": "integer",
                    "minimum": 1
                },
                "notes": {
                    "type": "string"
                },
                "policy_id": {
                    "description": "정책 ID (선택사항)",
                    "type": "string"
                },
                "product_id": {
                    "description": "제품 ID (선택사항)",
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                }
            }
        },
...
        "models.License": {
            "type": "object",
            "properties": {
                "active_devices": {
                    "description": "활성 디바이스 수",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "license_key": {
                    "type": "string"
                },
                "max_devices": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "policy_id": {
                    "type": "string"
                },
                "policy_name": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "status": {
                    "description": "active, revoked, expired",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.Pagination"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 토큰을 입력하세요. 형식: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}